Projects/AiReasoningLoop/pyproject.toml
[project]
name = "aireasoningloop"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "gradio>=6.4.0",
    "openai>=2.15.0",
    "python-dotenv>=1.2.1",
    "rich>=14.2.0",
]
